Remove bootstrap.js
Remove bootstrap.js

Removes dependency on bootstrap.js which is unused in the theme.
It can be introduced later if it later transpires that we need it.

Remove bootstrap.js and dependencies

Remove bootstrap.js

Removes dependency on bootstrap.js which is unused in the theme.
It can be introduced later if it later transpires that we need it.

Remove Bootstrap dependencies