Introduce output.html
The output.html file is generated by mega-menu-feed-processor for inclusion in other applications. Because this is managed outside of this repository - but developers still need a local copy - a function called set_path_to_mega_menu() is introduced with associated tests. This has been placed within a new include (tna-globals.php) which has been introduced to contain generic code that manages globals.