Move qunit tests from PhantomJS to Headless Chromium
PhantomJS is no longer maintained and I was unable to get the tests
to pass in a development environment running the latest Debian Linux.

Update grunt-contrib-qunit to the latest version, which uses
Puppeteer to drive the latest Chromium version instead of PhantomJS.

grunt-contrib-qunit 6 requires QUnit 2.2.0 or later, so while at it
also update the test files from QUnit 2.0.1 to 2.19.0.

See also:
* <https://github.com/gruntjs/grunt-contrib-qunit>.
* <https://pptr.dev/#?product=Puppeteer&version=v9.0.0&show=api-puppeteerlaunchoptions>